backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
httpx>=0.24
python-dotenv>=1.0
numpy>=1.25
pytest>=7.4
google-generativeai>=0.1.0
